Update index.php
- fixed bug in better security function
- improved function for password changing
- added keyword and description for 404 pages (new installations only)
- improved function for installing themes and plugins
- fixed bug with function for deleting files and folders

+++ index.php
@@ -1,7 +1,7 @@
 <?php // WonderCMS - MIT license: wondercms.com/license
 
 session_start();
-define('version', '2.5.0');
+define('version', '2.5.1');
 mb_internal_encoding('UTF-8');
 
 class wCMS
@@ -116,23 +116,24 @@ class wCMS
 
 	private static function betterSecurityAction()
 	{
-		if (! wCMS::$loggedIn && ! isset($_POST['betterSecurity'])) {
-			return;
-		}
-		if (hash_equals($_POST['token'], wCMS::generateToken()) && $_POST['betterSecurity'] == 'on') {
-			$contents = wCMS::getExternalFile('https://raw.githubusercontent.com/robiso/wondercms/master/.htaccess-ultimate');
-			if ($contents) {
-				file_put_contents('.htaccess', trim($contents));
+		if (wCMS::$loggedIn && isset($_POST['betterSecurity']) && isset($_POST['token'])) {
+			if (hash_equals($_POST['token'], wCMS::generateToken())) {
+				if ($_POST['betterSecurity'] == 'on') {
+					$contents = wCMS::getExternalFile('https://raw.githubusercontent.com/robiso/wondercms/master/.htaccess-ultimate');
+					if ($contents) {
+						file_put_contents('.htaccess', trim($contents));
+					}
+					wCMS::alert('success', 'Better security turned ON.');
+					wCMS::redirect();
+				} elseif ($_POST['betterSecurity'] == 'off') {
+					$contents = wCMS::getExternalFile('https://raw.githubusercontent.com/robiso/wondercms/master/.htaccess');
+					if ($contents) {
+						file_put_contents('.htaccess', trim($contents));
+					}
+					wCMS::alert('success', 'Better security turned OFF.');
+					wCMS::redirect();
+				}
 			}
-			wCMS::alert('success', 'Better security turned ON.');
-			wCMS::redirect();
-		} elseif (hash_equals($_POST['token'], wCMS::generateToken()) && $_POST['betterSecurity'] == 'off') {
-			$contents = wCMS::getExternalFile('https://raw.githubusercontent.com/robiso/wondercms/master/.htaccess');
-			if ($contents) {
-				file_put_contents('.htaccess', trim($contents));
-			}
-			wCMS::alert('success', 'Better security turned OFF.');
-			wCMS::redirect();
 		}
 	}
 
@@ -144,21 +145,20 @@ class wCMS
 
 	private static function changePasswordAction()
 	{
-		if (! wCMS::$loggedIn || ! isset($_POST['old_password']) || ! isset($_POST['new_password'])) {
-			return;
-		}
-		if ($_SESSION['token'] === $_POST['token'] && hash_equals($_POST['token'], wCMS::generateToken())) {
-			if (! password_verify($_POST['old_password'], wCMS::get('config', 'password'))) {
-				wCMS::alert('danger', 'Wrong password.');
+		if (wCMS::$loggedIn && isset($_POST['old_password']) && isset($_POST['new_password'])) {
+			if ($_SESSION['token'] === $_POST['token'] && hash_equals($_POST['token'], wCMS::generateToken())) {
+				if (! password_verify($_POST['old_password'], wCMS::get('config', 'password'))) {
+					wCMS::alert('danger', 'Wrong password.');
+					wCMS::redirect();
+				}
+				if (strlen($_POST['new_password']) < 4) {
+					wCMS::alert('danger', 'Password must be longer than 4 characters.');
+					wCMS::redirect();
+				}
+				wCMS::set('config', 'password', password_hash($_POST['new_password'], PASSWORD_DEFAULT));
+				wCMS::alert('success', 'Password changed.');
 				wCMS::redirect();
 			}
-			if (strlen($_POST['new_password']) < 4) {
-				wCMS::alert('danger', 'Password must be longer than 4 characters.');
-				wCMS::redirect();
-			}
-			wCMS::set('config', 'password', password_hash($_POST['new_password'], PASSWORD_DEFAULT));
-			wCMS::alert('success', 'Password changed.');
-			wCMS::redirect();
 		}
 	}
 
@@ -191,8 +191,8 @@ class wCMS
 			'pages' => [
 				'404' => [
 					'title' => '404',
-					'keywords' => '',
-					'description' => '',
+					'keywords' => '404',
+					'description' => '404',
 					'content' => '<h1>Sorry, page not found. :(</h1>'
 				],
 				'home' => [
@@ -433,9 +433,14 @@ EOT;
 			return;
 		}
 		if (hash_equals($_POST['token'], wCMS::generateToken())) {
-			$installLocation = trim(strtolower($_POST['installLocation']));
-			$addonURL = $_POST['addonURL'];
-			$validPaths = array("themes", "plugins");
+			if (isset($_POST['installLocation'])) {
+				$installLocation = trim(strtolower($_POST['installLocation']));
+				$addonURL = $_POST['addonURL'];
+				$validPaths = array("themes", "plugins");
+			} else {
+				wCMS::alert('danger', 'Choose between theme or plugin.');
+				wCMS::redirect();
+			}
 			if (in_array($installLocation, $validPaths) && ! empty($addonURL)) {
 				$zipFile = __DIR__ . '/files/ZIPFromURL.zip';
 				$zipResource = fopen($zipFile, "w");
@@ -456,9 +461,6 @@ EOT;
 				$zip->close();
 				wCMS::recursiveDelete(__DIR__ . '/files/ZIPFromURL.zip');
 				wCMS::alert('success', 'Installed successfully.');
-				wCMS::redirect();
-			} elseif (empty($installLocation)) {
-				wCMS::alert('danger', 'Choose between theme or plugin.');
 				wCMS::redirect();
 			} else {
 				wCMS::alert('danger', 'Enter URL to ZIP file.');
@@ -612,7 +614,9 @@ EOT;
 			foreach ($list as $dir) {
 				wCMS::recursiveDelete($dir);
 			}
-			rmdir($file);
+			if (file_exists($file)) {
+				rmdir($file);
+			}
 		} elseif (is_file($file)) {
 			unlink($file);
 		}
